# src/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CreateBookRequest:
    properties:
      author_ids:
        items:
          type: string
        type: array
      category_id:
        type: string
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      original_price:
        type: number
      page_count:
        type: integer
      price:
        type: number
      rating_average:
        type: number
      review_count:
        type: integer
      secondary_id:
        type: string
      seller_id:
        type: string
      stock:
        type: integer
    required:
      - author_ids
      - category_id
      - name
      - price
      - seller_id
    type: object
...
  dtos.UpdateBookRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      original_price:
        type: number
      page_count:
        type: integer
      price:
        type: number
      rating_average:
        type: number
      review_count:
        type: integer
      stock:
        type: integer
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      error: {}
      headers: {}
      message:
        type: string
      pagination: {}
      status_code:
        type: integer
    type: object
host: localhost:5000
info:
  contact:
    email: tntuan0910@gmail.com
    name: tuan.tran
  description: This is the API documentation for the Rainbow API.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Rainbow API
  version: '1.0'
paths:
...
  /books:
    get:
      consumes:
        - application/json
      description: Fetch a list of books
      parameters:
        - description: Page number
          in: query
          name: page
          type: integer
        - description: Number of items per page
          in: query
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get Books
      tags:
        - Book
  /books/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a book by its ID
      parameters:
        - description: Book ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/response.APIResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Delete Book
      tags:
        - Book
    get:
      consumes:
        - application/json
      description: Fetch a book by its ID
      parameters:
        - description: Book ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get Book
      tags:
        - Book
    patch:
      consumes:
        - application/json
      description: Update a book by its ID
      parameters:
        - description: Update Book Request
          in: body
          name: req
          required: true
          schema:
            $ref: '#/definitions/dtos.UpdateBookRequest'
      produces:
        - application/json
      responses:
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/response.APIResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Update Book
      tags:
        - Book
    post:
      consumes:
        - application/json
      description: Create a book
      parameters:
        - description: Create Book Request
          in: body
          name: req
          required: true
          schema:
            $ref: '#/definitions/dtos.CreateBookRequest'
      produces:
        - application/json
      responses:
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/response.APIResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Create Book
      tags:
        - Book
...
swagger: '2.0'